read/write enum arrays
