+++ src/NCDatasets.jl
@@ -64,6 +64,7 @@ include("errorhandling.jl")
 include("lock.jl")
 include("netcdf_c.jl")
 include("compound_types.jl")
+include("enum_types.jl")
 include("dataset.jl")
 include("attributes.jl")
 include("dimensions.jl")
